iot-sensor-client/requirements.txt
paho-mqtt>=2.0
PyYAML>=6.0
smbus2>=0.4
